Organizing components Refactoring Website with Navbar component
