--- frontend/src/components/News.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function News(){
-    return(
-        <div>
-            News Stories
-        </div>
-    )
-}
-
-export default News;
