--- frontend/src/components/Website/Website.css
@@ -1,51 +1,0 @@
-.flex_1 {
-  display: flex;
-  width: 35%;
-}
-
-.flex_2 {
-  display: flex;
-  width: 60%;
-}
-
-.pageContainer {
-  height: 100%;
-  /* overflow-y: hidden; */
-}
-
-ol.breadcrumb {
-  margin: 0;
-}
-
-.breadcrumb {
-  width: 100%;
-}
-
-@media (max-width: 900px) {
-  .breadcrumb {
-    width: 910px;
-  }
-
-  .flexContainer {
-    width: 910px;
-  }
-}
-
-.pageContainer__flexCenter {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding-bottom: 15em;
-}
-
-.comingSoonContainer {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.loadingIndicator {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
